split in ROS and nonROS the simulator
