+++ sim_mujoco/sim_mujoco/mujoco_node.py
@@ -12,12 +12,12 @@ from std_msgs.msg import Bool, Float64, Empty
 from tf2_ros import TransformBroadcaster
 from scipy.spatial.transform import Rotation as R
 
-import mujoco.viewer
 import mujoco as mj
 import numpy as np
 import time
 from threading import Lock
 import math 
+import mujoco_env_sim as mujoco_sim
 
 
 class MujocoNode(Node):
@@ -34,23 +34,13 @@ class MujocoNode(Node):
         self.publish_tf = self.get_parameter("publish_tf").get_parameter_value().bool_value
         self.initialization_done = False
         self.goal_pos = [0.0, 0.0]
-        self.contact_states = {'R_FOOT': False,
-                               'L_FOOT': False}
 
-        self.model = mj.MjModel.from_xml_path(mujoco_xml_path)
-        self.data = mj.MjData(self.model)
-
-        if self.visualize_mujoco is True:
-            self.get_logger().info("Start visualization!")
-            self.viewer = mujoco.viewer.launch_passive(self.model, self.data)
-
-        mj.mj_printModel(self.model, 'robot_information.txt')
+        self.biped = mujoco_sim.Biped(mujoco_xml_path, self.sim_time_sec)
 
         self.lock = Lock()
 
         self.time = time.time()
-        self.model.opt.timestep = self.sim_time_sec
-        self.dt = self.model.opt.timestep
+        self.dt = self.biped.get_sim_dt()
 
         self.accel_noise_density = 0.14 * 9.81/1000 # [m/s2 * sqrt(s)]
         self.accel_bias_random_walk = 0.0004 # [m/s2 / sqrt(s)]
@@ -63,41 +53,18 @@ class MujocoNode(Node):
         self.accel_bias = np.random.normal(0, self.accel_bias_random_walk * np.sqrt(100), 3)
         self.gyro_bias = np.random.normal(0, self.gyro_bias_random_walk * np.sqrt(100), 3)
 
-        self.name_joints = self.get_joint_names()
+        self.name_joints = self.biped.get_joint_names()
+        self.q_joints = self.biped.get_q_joints_dict()
 
-        self.q_joints = {}
-        for i in self.name_joints:
-            self.q_joints[i] = {
-                'timestamp': 0.0,
-                'actual_pos': 0.0,
-                'actual_vel': 0.0,
-                'actual_acc': 0.0,
-                'actual_effort': 0.0,
-                'desired_pos': 0.0,
-                'desired_vel': 0.0,
-                'feedforward_torque': 0.0,
-                'qfrc_actuator': 0.0,
-                'total_tau': 0.0,
-                'qfrc_passive': 0.0
-            }
+        self.name_actuators = self.biped.get_name_actuators()
+        self.q_actuator_addr = self.biped.get_q_actuator_addr()
+        self.q_pos_addr_joints = self.biped.get_q_pos_addr_joints()
 
-        self.name_actuators = []
-        for i in range(0, self.model.nu):  # skip root
-            self.name_actuators.append(mj.mj_id2name(self.model, mj.mjtObj.mjOBJ_ACTUATOR, i))
-
-        self.q_actuator_addr = {}
-        for name in self.name_actuators:
-            self.q_actuator_addr[name] = mj.mj_name2id(self.model, mj.mjtObj.mjOBJ_ACTUATOR, name)
-
-        self.q_pos_addr_joints = {}
-        for name in self.name_joints:
-            self.q_pos_addr_joints[name] = self.model.jnt_qposadr[mj.mj_name2id(
-                self.model, mj.mjtObj.mjOBJ_JOINT, name)]
-        
         self.counter = 0
         self.init(p=[0.0, 0.0, 0.0])
-        self.clock_pub = self.create_publisher(Clock, '/clock', 10)
 
+        # Publishers.
+        self.clock_pub = self.create_publisher(Clock, '/clock', 10)
         self.odometry_base_pub = self.create_publisher(Odometry, '~/odometry', 10)
         self.contact_right_pub = self.create_publisher(StampedBool, '~/contact_foot_right', 10)
         self.contact_left_pub = self.create_publisher(StampedBool, '~/contact_foot_left', 10)
@@ -113,6 +80,7 @@ class MujocoNode(Node):
         self.fake_vicon_pub = self.create_publisher(PoseStamped, '~/fake_vicon', 10)
         self.tf_broadcaster = TransformBroadcaster(self)
 
+        # Subscribers.
         self.joint_traj_sub = self.create_subscription(JointTrajectory, 'joint_trajectory', self.joint_traj_cb, 10)
         self.joint_traj_msg = None
         self.initial_pose_sub = self.create_subscription(PoseWithCovarianceStamped, 'initialpose', self.init_cb, 10)
@@ -122,7 +90,7 @@ class MujocoNode(Node):
         self.step_sim_sub = self.create_subscription(Float64, "~/step", self.step_cb, 1)
         self.pause_sim_sub = self.create_subscription(Bool, "~/pause", self.pause_cb, 1)
 
-        self.previous_q_vel = np.zeros(self.model.nv)
+        self.previous_q_vel = np.zeros(self.biped.get_nv())
 
         self.timer = self.create_timer(self.dt*2, self.timer_cb, clock=rclpy.clock.Clock(clock_type=rclpy.clock.ClockType.STEADY_TIME))
 
@@ -137,27 +105,10 @@ class MujocoNode(Node):
             self.init([p.x, p.y, p.z], q=[q.w, q.x, q.y, q.z])
 
     def init(self, p, q=[1.0, 0.0, 0.0, 0.0]):
-        self.model.eq_data[0][0] = p[0]
-        self.model.eq_data[0][1] = p[1]
-        self.model.eq_data[0][2] = 1.0 # one meter above gnd
-
-        self.data.qpos = [0.0] * self.model.nq
-        self.data.qpos[3] = q[0]
-        self.data.qpos[4] = q[1]
-        self.data.qpos[5] = q[2]
-        self.data.qpos[6] = q[3]
-
-        self.data.qvel = [0.0]* self.model.nv
-
-        self.data.qpos[0] = p[0]
-        self.data.qpos[1] = p[1]
-        self.data.qpos[2] = self.model.eq_data[0][2]
-
-        self.data.eq_active[0] = 1
-        mj.mj_step(self.model, self.data)
+        self.biped.init(p, q)
+        self.get_logger().info("initialize")
         self.initialization_done = False
         self.initialization_timeout = 0.2
-        self.get_logger().info("initialize")
 
     def timer_cb(self):
         with self.lock:
@@ -179,36 +130,28 @@ class MujocoNode(Node):
             self.paused = msg.data
 
     def stop_viz(self):
-        self.viewer.close()
+        self.biped.viewer.close()
 
     def step(self):
         if not self.initialization_done:
             msg_stop_controller = Bool()
             msg_stop_controller.data = (self.initialization_timeout > 0)
             self.stop_pub.publish(msg_stop_controller)
-
-            self.model.eq_data[0][2] -= 0.5 * self.dt
+            self.biped.lower_robot(step=self.dt)
             self.initialization_timeout -= self.dt
-        self.read_contact_states()
-        if self.contact_states['R_FOOT'] or self.contact_states['L_FOOT']:
+
+        contact_states = self.read_contact_states()
+        if contact_states['R_FOOT'] or contact_states['L_FOOT']:
             if not self.initialization_done:
                 self.get_logger().info("init done")
                 self.initialization_done = True
-                self.data.qvel = [0.0]* self.model.nv
-                self.model.eq_active0 = 0 # let go of the robot
-                self.data.eq_active[0] = 0 # let go of the robot
+                self.biped.zero_the_velocities()
+                self.biped.let_go_of_robot()
 
-        for _ in range(2):
-            self.run_joint_controllers()
-            self.ankle_foot_spring('L_ANKLE')
-            self.ankle_foot_spring('R_ANKLE')
-            mj.mj_step(self.model, self.data)
-            if self.visualize_mujoco is True:
-                if self.viewer.is_running():
-                    self.viewer.sync()
-            self.time += self.dt
-            self.counter += 1
+        is_valid_traj_msg = False if self.joint_traj_msg is None else True
+        qpos, qvel = self.biped.step(is_valid_traj_msg)
 
+        # ROS publishers.
         clock_msg = Clock()
         clock_msg.clock.sec = int(self.time)
         clock_msg.clock.nanosec = int((self.time - clock_msg.clock.sec) * 1e9)
@@ -219,23 +162,23 @@ class MujocoNode(Node):
         msg_odom.header.stamp.nanosec = int((self.time - clock_msg.clock.sec) * 1e9)
         msg_odom.header.frame_id = 'odom'
         msg_odom.child_frame_id = 'base_link'
-        msg_odom.pose.pose.position.x = self.data.qpos[0]
-        msg_odom.pose.pose.position.y = self.data.qpos[1]
-        msg_odom.pose.pose.position.z = self.data.qpos[2]
-        msg_odom.pose.pose.orientation.w = self.data.qpos[3]
-        msg_odom.pose.pose.orientation.x = self.data.qpos[4]
-        msg_odom.pose.pose.orientation.y = self.data.qpos[5]
-        msg_odom.pose.pose.orientation.z = self.data.qpos[6]
+        msg_odom.pose.pose.position.x = qpos[0]
+        msg_odom.pose.pose.position.y = qpos[1]
+        msg_odom.pose.pose.position.z = qpos[2]
+        msg_odom.pose.pose.orientation.w = qpos[3]
+        msg_odom.pose.pose.orientation.x = qpos[4]
+        msg_odom.pose.pose.orientation.y = qpos[5]
+        msg_odom.pose.pose.orientation.z = qpos[6]
         q = msg_odom.pose.pose.orientation
 
         R_b_to_I = R.from_quat([q.x, q.y, q.z, q.w]).as_matrix()
-        v_b = R_b_to_I.T @ self.data.qvel[0:3] # linear vel is in inertial frame
+        v_b = R_b_to_I.T @ qvel[0:3] # linear vel is in inertial frame
         msg_odom.twist.twist.linear.x = v_b[0]
         msg_odom.twist.twist.linear.y = v_b[1]
         msg_odom.twist.twist.linear.z = v_b[2]
-        msg_odom.twist.twist.angular.x = self.data.qvel[3] # angular vel is in body frame
-        msg_odom.twist.twist.angular.y = self.data.qvel[4]
-        msg_odom.twist.twist.angular.z = self.data.qvel[5]
+        msg_odom.twist.twist.angular.x = qvel[3] # angular vel is in body frame
+        msg_odom.twist.twist.angular.y = qvel[4]
+        msg_odom.twist.twist.angular.z = qvel[5]
         self.odometry_base_pub.publish(msg_odom)
 
         if self.publish_tf:
@@ -248,15 +191,15 @@ class MujocoNode(Node):
             t.transform.rotation = msg_odom.pose.pose.orientation
             self.tf_broadcaster.sendTransform(t)
 
-        self.read_contact_states()
+        contact_states = self.read_contact_states()
         msg_contact_right = StampedBool()
         msg_contact_left = StampedBool()
         msg_contact_right.header.stamp.sec = int(self.time)
         msg_contact_right.header.stamp.nanosec = int((self.time - clock_msg.clock.sec) * 1e9)
-        msg_contact_right.data = self.contact_states['R_FOOT']
+        msg_contact_right.data = contact_states['R_FOOT']
         msg_contact_left.header.stamp.sec = int(self.time)
         msg_contact_left.header.stamp.nanosec = int((self.time - clock_msg.clock.sec) * 1e9)
-        msg_contact_left.data = self.contact_states['L_FOOT']
+        msg_contact_left.data = contact_states['L_FOOT']
         self.contact_right_pub.publish(msg_contact_right)
         self.contact_left_pub.publish(msg_contact_left)
 
@@ -297,12 +240,10 @@ class MujocoNode(Node):
             msg_qfrc_passive.effort.append(value['qfrc_passive'])
         self.qfrc_passive_pub.publish(msg_qfrc_passive)
 
-        gyro_id = mj.mj_name2id(self.model, mj.mjtObj.mjOBJ_SENSOR, "gyro")
-        accel_id = mj.mj_name2id(self.model, mj.mjtObj.mjOBJ_SENSOR, "accelerometer")
-        accel = self.data.sensordata[self.model.sensor_adr[accel_id]:self.model.sensor_adr[accel_id] + 3]
+        gyro = self.biped.get_gyro_data()
+        accel = self.biped.get_accel_data()
         self.accel_bias += np.random.normal(0, self.accel_bias_noise_std, 3)
         # accel += self.accel_bias + np.random.normal(0, self.accel_noise_std, 3)
-        gyro = self.data.sensordata[self.model.sensor_adr[gyro_id]:self.model.sensor_adr[gyro_id] + 3]
         self.gyro_bias += np.random.normal(0, self.gyro_bias_noise_std, 3)
         # gyro += self.gyro_bias + np.random.normal(0, self.gyro_noise_std, 3)
         if not self.initialization_done:
@@ -325,120 +266,18 @@ class MujocoNode(Node):
         msg_fake_vicon.header.stamp.sec = int(self.time)
         msg_fake_vicon.header.stamp.nanosec = int((self.time - clock_msg.clock.sec) * 1e9)
         msg_fake_vicon.header.frame_id = 'world'
-        msg_fake_vicon.pose.position.x = self.data.qpos[0]
-        msg_fake_vicon.pose.position.y = self.data.qpos[1]
-        msg_fake_vicon.pose.position.z = self.data.qpos[2]
-        msg_fake_vicon.pose.orientation.w = self.data.qpos[3]
-        msg_fake_vicon.pose.orientation.x = self.data.qpos[4]
-        msg_fake_vicon.pose.orientation.y = self.data.qpos[5]
-        msg_fake_vicon.pose.orientation.z = self.data.qpos[6]
+        msg_fake_vicon.pose.position.x = qpos[0]
+        msg_fake_vicon.pose.position.y = qpos[1]
+        msg_fake_vicon.pose.position.z = qpos[2]
+        msg_fake_vicon.pose.orientation.w = qpos[3]
+        msg_fake_vicon.pose.orientation.x = qpos[4]
+        msg_fake_vicon.pose.orientation.y = qpos[5]
+        msg_fake_vicon.pose.orientation.z = qpos[6]
         self.fake_vicon_pub.publish(msg_fake_vicon)
-
-    def run_joint_controllers(self):
-        if self.joint_traj_msg is None:
-            return
-        for key, value in self.q_joints.items():
-            id_joint_mj = mj.mj_name2id(self.model, mj.mjtObj.mjOBJ_JOINT, key)
-            value['actual_pos'] = self.data.qpos[self.model.jnt_qposadr[id_joint_mj]]
-            value['actual_vel'] = self.data.qvel[self.model.jnt_dofadr[id_joint_mj]]
-            actual_acc = (self.data.qvel[self.model.jnt_dofadr[id_joint_mj]] - self.previous_q_vel[self.model.jnt_dofadr[id_joint_mj]])/self.dt
-            value['actual_acc'] = actual_acc
-            value['actual_effort'] = self.data.qfrc_actuator[self.model.jnt_dofadr[id_joint_mj]] + self.data.qfrc_applied[self.model.jnt_dofadr[id_joint_mj]]
-            if key in self.joint_traj_msg.joint_names:
-                id_joint_msg = self.joint_traj_msg.joint_names.index(key)
-                value['desired_pos'] = self.joint_traj_msg.points[0].positions[id_joint_msg]
-                if self.joint_traj_msg.points[0].velocities:
-                    value['desired_vel'] = self.joint_traj_msg.points[0].velocities[id_joint_msg]
-                if self.joint_traj_msg.points[0].effort:
-                    if math.isnan(self.joint_traj_msg.points[0].effort[id_joint_msg]) is False:
-                        value['feedforward_torque'] = self.joint_traj_msg.points[0].effort[id_joint_msg]
-                    else:
-                        value['feedforward_torque'] = 0.0
-        self.previous_q_vel = self.data.qvel.copy()
-
-        kp_moteus = 240.0
-        Kp = (kp_moteus/(2*math.pi)) * np.ones(self.model.njnt - 1) # exclude root
-
-        i = 0
-        for key, value in self.q_joints.items():
-            if key != 'L_ANKLE' and key != 'R_ANKLE':
-                # Q: why are we not using a PD controller
-                error = value['actual_pos'] - value['desired_pos']
-                actuators_torque = - Kp[i] * error
-                actuators_vel = value['desired_vel']
-                feedforward_torque = value['feedforward_torque']
-                self.data.ctrl[self.q_actuator_addr[str(key)]] = actuators_torque
-                self.data.ctrl[self.q_actuator_addr[str(key) + "_VEL"]] = actuators_vel
-                id_joint_mj = mj.mj_name2id(self.model, mj.mjtObj.mjOBJ_JOINT, key)
-
-                self.data.qfrc_applied[self.model.jnt_dofadr[id_joint_mj]] = feedforward_torque
-                self.q_joints[key]['qfrc_actuator'] = self.data.qfrc_actuator[self.model.jnt_dofadr[id_joint_mj]]
-
-                self.q_joints[key]['total_tau'] = self.data.qfrc_passive[self.model.jnt_dofadr[id_joint_mj]] + \
-                                                    self.data.qfrc_actuator[self.model.jnt_dofadr[id_joint_mj]] + \
-                                                    self.data.qfrc_applied[self.model.jnt_dofadr[id_joint_mj]]
-            i = i + 1
-
-
-    # def stop_controller(self, actuator_name):
-    #     idx_act = mj.mj_name2id(
-    #         self.model, mj.mjtObj.mjOBJ_ACTUATOR, actuator_name)
-    #     self.data.ctrl[idx_act] = 0.0
 
     def joint_traj_cb(self, msg):
         with self.lock:
             self.joint_traj_msg = msg
-
-    def read_contact_states(self):
-        self.contact_states['R_FOOT'] = False
-        self.contact_states['L_FOOT'] = False
-
-        geom1_list = []
-        geom2_list = []
-        for i in range(self.data.ncon):
-            contact = self.data.contact[i]
-
-            name_geom1 = mj.mj_id2name(
-                self.model, mj.mjtObj.mjOBJ_GEOM, contact.geom1)
-            name_geom2 = mj.mj_id2name(
-                self.model, mj.mjtObj.mjOBJ_GEOM, contact.geom2)
-            geom1_list.append(name_geom1)
-            geom2_list.append(name_geom2)
-
-        if self.data.ncon != 0:
-            if 'L_FOOT' in geom2_list:
-                first_entry_idx = geom2_list.index('L_FOOT')
-                if geom1_list[first_entry_idx] == 'floor':
-                    self.contact_states['L_FOOT'] = True
-
-            if 'R_FOOT' in geom2_list:
-                first_entry_idx = geom2_list.index('R_FOOT')
-                if geom1_list[first_entry_idx] == 'floor':
-                    self.contact_states['R_FOOT'] = True
-
-    def get_joint_names(self):
-        self.name_joints = []
-        for i in range(1, self.model.njnt):  # skip root
-            self.name_joints.append(mj.mj_id2name(
-                self.model, mj.mjtObj.mjOBJ_JOINT, i))
-        return self.name_joints
-
-
-    def ankle_foot_spring(self, foot_joint):
-        'ankle modelled as a spring damped system'
-        K = 0.005
-        offset = 0.5
-        pitch_error_foot = self.data.qpos[self.q_pos_addr_joints[foot_joint]] - offset
-        pitch_torque_setpt = - K * pitch_error_foot
-
-        idx_act = mj.mj_name2id(
-            self.model, mj.mjtObj.mjOBJ_ACTUATOR, foot_joint)
-        vel_actuator_name = str(foot_joint) + str('_VEL')
-        idx_vel_act = mj.mj_name2id(
-            self.model, mj.mjtObj.mjOBJ_ACTUATOR, vel_actuator_name)
-
-        self.data.ctrl[idx_act] = pitch_torque_setpt
-        self.data.ctrl[idx_vel_act] = 0.0
 
 def main(args=None):
     rclpy.init(args=args)
